+++ car_parking/src/routes/admin.py
@@ -4,9 +4,9 @@ from sqlalchemy.orm import Session
 
 from ..database.db import get_db
 from ..database.models import User
-from ..repository import users as repository_admin
+from ..repository import users as repository_users, admin as repository_admin
 from ..services.auth import service_auth
-from ..schemas.users import UserResponce
+from ..schemas.users import UserResponce, UserRoleUpdate
 from ..services import (
     roles as service_roles,
     logout as service_logout,
@@ -25,8 +25,7 @@ allowd_operation_by_admin = service_roles.RoleRights(["admin"])
 @router.get('/',
             status_code=status.HTTP_200_OK,
             dependencies=[Depends(service_logout.logout_dependency), 
-                          Depends(allowd_operation),
-                          Depends(service_banned.banned_dependency)]
+                          Depends(allowd_operation_by_admin)]
             )
 async def get_all_usernames(current_user: User = Depends(service_auth.get_current_user),
                             db: Session = Depends(get_db)):
@@ -48,7 +47,7 @@ async def get_all_usernames(current_user: User = Depends(service_auth.get_curren
                         dependencies=[Depends(service_logout.logout_dependency), 
                                       Depends(allowd_operation_by_admin)],
                         )
-async def update_banned_status(user_id:str,
+async def ban_user(user_id:str,
                                 credentials: HTTPAuthorizationCredentials = Security(security), 
                                 current_user: User = Depends(service_auth.get_current_user),
                                 db: Session = Depends(get_db)):
@@ -82,7 +81,7 @@ async def update_banned_status(user_id:str,
                         dependencies=[Depends(service_logout.logout_dependency), 
                                       Depends(allowd_operation_by_admin)],
                         )
-async def update_unbanned_status(user_id:str,
+async def unban_user(user_id:str,
                                 credentials: HTTPAuthorizationCredentials = Security(security), 
                                 current_user: User = Depends(service_auth.get_current_user),
                                 db: Session = Depends(get_db)):
@@ -139,3 +138,42 @@ async def delete_user(user_id: int, current_user: User = Depends(service_auth.ge
 
     await repository_users.delete_user(user_id, db)
     return {"message": f"User successfully deleted"}
+
+
+@router.patch('/change_role/{user_id}', response_model=UserResponce, 
+                                        status_code=status.HTTP_202_ACCEPTED,
+                                        dependencies=[Depends(service_logout.logout_dependency), 
+                                                      Depends(allowd_operation_by_admin)])
+async def change_user_role(
+    user_id: int,
+    body: UserRoleUpdate,
+    current_user: User = Depends(service_auth.get_current_user),
+    db: Session = Depends(get_db)
+):
+    """
+    The change_user_role function allows an admin to change the role of a user.
+    
+    :param user_id: int: Fetch the user by id from the database
+    :param body: UserRoleUpdate: Get the new role from the request body
+    :param current_user: User: Get the current user from the database
+    :param db: Session: Pass the database session to the function
+    :return: A user object with udated role
+    :doc-author: Trelent
+    """
+    if current_user.role != "admin":
+        raise HTTPException(status_code=403, detail="Permission denied. Only admin can change roles.")
+    user = await repository_users.get_user_by_id(user_id, db)
+    if not user:
+        raise HTTPException(status_code=404, detail="User not found")
+    if current_user.id == user_id:
+        raise HTTPException(status_code=403, detail="Permission denied. Own role cannot be changed.")
+    if user.id == 1:
+        raise HTTPException(status_code=403, detail="Permission denied.Superadmin role cannot be changed.")
+    if user.role == "admin" and current_user.id != 1:
+        raise HTTPException(status_code=403, detail="Permission denied.Admin role can be changed only by Superadmin (id=1).")
+
+    if body.role in ['admin', 'user']:
+        await repository_admin.change_user_role(user, body, db)
+        return user
+    else:
+        raise HTTPException(status_code=400, detail="Invalid role provided")
