+++ car_parking/src/routes/auth.py
@@ -215,45 +215,6 @@ async def reset_password(body: schema_users.ChangePassword, token: str, db: Sess
     return {"detail": "User's password was changed succesfully"}
 
 
-@router.patch('/change_role/{user_id}', response_model=schema_users.UserResponce, 
-                                        status_code=status.HTTP_202_ACCEPTED,
-                                        dependencies=[Depends(service_logout.logout_dependency), 
-                                                      Depends(allowd_operation_by_admin)])
-async def change_user_role(
-    user_id: int,
-    body: schema_users.UserRoleUpdate,
-    current_user: User = Depends(service_auth.get_current_user),
-    db: Session = Depends(get_db)
-):
-    """
-    The change_user_role function allows an admin to change the role of a user.
-    
-    :param user_id: int: Fetch the user by id from the database
-    :param body: UserRoleUpdate: Get the new role from the request body
-    :param current_user: User: Get the current user from the database
-    :param db: Session: Pass the database session to the function
-    :return: A user object with udated role
-    :doc-author: Trelent
-    """
-    if current_user.role != "admin":
-        raise HTTPException(status_code=403, detail="Permission denied. Only admin can change roles.")
-    user = await repository_users.get_user_by_id(user_id, db)
-    if not user:
-        raise HTTPException(status_code=404, detail="User not found")
-    if current_user.id == user_id:
-        raise HTTPException(status_code=403, detail="Permission denied. Own role cannot be changed.")
-    if user.id == 1:
-        raise HTTPException(status_code=403, detail="Permission denied.Superadmin role cannot be changed.")
-    if user.role == "admin" and current_user.id != 1:
-        raise HTTPException(status_code=403, detail="Permission denied.Admin role can be changed only by Superadmin (id=1).")
-
-    if body.role in ['admin', 'moderator', 'user']:
-        await repository_users.change_user_role(user, body, db)
-        return user
-    else:
-        raise HTTPException(status_code=400, detail="Invalid role provided")
-    
-
 @router.get('/logout',
             status_code=status.HTTP_200_OK,
             dependencies=[Depends(service_logout.logout_dependency),
